Revert "componentized page elements"
This reverts commit 686e9b5649cdb1769d81ead340d3c364392f7c6d.
